Add token-based authentication, implement corresponding tests

+++ controllers/blogs.js
@@ -1,14 +1,17 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog
+        .find({}).populate('user', { username: 1, name: 1 })
     response.json(blogs.map(blog => blog.toJSON()))
 })
 
 blogsRouter.get('/:id', async (request, response) => {
     const blog = await Blog.findById(Request.params.id)
-    
+
     if (blog) {
         response.json(blog.toJSON())
     } else {
@@ -17,19 +20,39 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response) => {
+    const body = request.body
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    const user = await User.findById(decodedToken.id)
+
     const blog = new Blog({
         title: request.body.title,
         author: request.body.author,
-        url: request.body.author,
-        likes: request.body.likes
+        url: request.body.url,
+        likes: request.body.likes,
+        user: user._id
     })
+
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog)
+    await user.save()
     response.status(201).json(savedBlog.toJSON())
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const blog = await Blog.findById(request.params.id)
+    if (!blog) {
+        response.status(404).json({ error: 'the blog is not found' })
+    }
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if (blog.user.toString() === decodedToken.id) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+    } else {
+        response.status(401).json({ error: 'unauthorized' })
+    }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
